Move Lightning Address generation into wallet CRUD
