+++ lnbits/core/crud/wallets.py
@@ -1,15 +1,76 @@
 from datetime import datetime, timezone
+from random import SystemRandom
 from time import time
 from typing import Any
 from uuid import uuid4
 
+from sqlalchemy.exc import OperationalError
+
 from lnbits.core.db import db
 from lnbits.core.models.wallets import BaseWallet, WalletsFilters, WalletType
-from lnbits.db import Connection, Filters, Page
+from lnbits.db import Connection, Database, Filters, Page
 from lnbits.settings import settings
 from lnbits.utils.cache import cache
 
 from ..models import Wallet
+
+_RANDOM = SystemRandom()
+_LEGACY_LNURLP_DB = Database("ext_lnurlp")
+
+_PARTICIPLES = [
+    "asking",
+    "blazing",
+    "bouncing",
+    "charging",
+    "climbing",
+    "dancing",
+    "drifting",
+    "flying",
+    "glowing",
+    "hopping",
+    "jumping",
+    "laughing",
+    "leaping",
+    "racing",
+    "rising",
+    "running",
+    "shining",
+    "singing",
+    "skipping",
+    "sparking",
+    "spinning",
+    "sprinting",
+    "twirling",
+    "wandering",
+    "zipping",
+]
+
+_NOUNS = [
+    "anchor",
+    "beacon",
+    "bolt",
+    "comet",
+    "ember",
+    "falcon",
+    "flash",
+    "garden",
+    "harbor",
+    "lantern",
+    "market",
+    "meadow",
+    "meteor",
+    "orbit",
+    "phoenix",
+    "rabbit",
+    "rocket",
+    "signal",
+    "spark",
+    "summit",
+    "thunder",
+    "wallet",
+    "wave",
+    "zephyr",
+]
 
 
 async def create_wallet(
@@ -32,8 +93,6 @@ async def create_wallet(
         currency=settings.lnbits_default_accounting_currency or "USD",
     )
     if settings.lnbits_enable_wallet_lightning_addresses and wallet.is_lightning_wallet:
-        from ..services.lightning_address import generate_lightning_address_local_part
-
         wallet.lightning_address = await generate_lightning_address_local_part(conn)
 
     await (conn or db).insert("wallets", wallet)
@@ -241,6 +300,43 @@ async def wallet_lightning_address_exists(
         {"lightning_address": local_part},
     )
     return bool(row)
+
+
+def _generate_local_part() -> str:
+    participle = _RANDOM.choice(_PARTICIPLES)
+    noun = _RANDOM.choice(_NOUNS)
+    suffix = _RANDOM.randint(0, 999)
+    return f"{participle}.{noun}.{suffix:03d}"
+
+
+async def legacy_lnurlp_address_exists(local_part: str) -> bool:
+    try:
+        row: Any = await _LEGACY_LNURLP_DB.fetchone(
+            """
+            SELECT 1 FROM lnurlp.pay_links
+            WHERE username = :username
+            LIMIT 1
+            """,
+            {"username": local_part},
+        )
+        return row is not None
+    except OperationalError:
+        return False
+    except Exception:
+        return False
+
+
+async def generate_lightning_address_local_part(
+    conn: Connection | None = None,
+) -> str:
+    for _ in range(100):
+        local_part = _generate_local_part()
+        if await wallet_lightning_address_exists(local_part, conn):
+            continue
+        if await legacy_lnurlp_address_exists(local_part):
+            continue
+        return local_part
+    raise ValueError("Could not generate a unique wallet lightning address.")
 
 
 async def wallet_lightning_address_exists_for_other_wallet(
